# packages/backend/src/openapi/spec.yaml
openapi: 3.0.3
info:
  title: Jyotish API
  description: |
    Vedic Astrology API for birth chart calculation and AI-powered readings.

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    API requests are rate limited. Check response headers for limit information.
  version: 1.0.0
  contact:
    name: Jyotish Support
  license:
    name: MIT
servers:
  - url: /
    description: Current server (relative URL)
  - url: http://localhost:3001
    description: Local development
  - url: http://192.168.1.225:3001
    description: LAN development
...
paths:
...
  /api/chart/{id}:
    get:
      summary: Get a saved chart
      operationId: getChart
      tags:
        - Chart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Chart ID
      responses:
        '200':
          description: Chart retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChartResponse'
        '404':
          description: Chart not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/reading/categories:
    get:
      summary: Get available reading categories
      operationId: getReadingCategories
      tags:
        - Reading
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: User ID to check access levels
      responses:
        '200':
          description: List of available categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriesResponse'
  /api/reading/questions/remaining:
    get:
      summary: Get remaining questions for a user
      operationId: getRemainingQuestions
      tags:
        - Reading
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: User ID (optional, defaults to anonymous)
      responses:
        '200':
          description: Question quota information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionQuotaResponse'
  /api/reading/{chartId}/summary:
    get:
      summary: Get AI-generated summary reading
      operationId: getSummaryReading
      tags:
        - Reading
      parameters:
        - name: chartId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Chart ID
        - name: provider
          in: query
          required: false
          schema:
            type: string
            enum: [claude, deepseek, gemini, openai]
          description: Preferred AI provider (optional, uses configured default if not specified)
      responses:
        '200':
          description: Summary reading generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingResponse'
        '404':
          description: Chart not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: AI service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/reading/{chartId}/{category}:
    get:
      summary: Get category-specific reading
      operationId: getCategoryReading
      tags:
        - Reading
      parameters:
        - name: chartId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Chart ID
        - name: category
          in: path
          required: true
          schema:
            type: string
            enum: [summary, love, career, finances, health, timeline]
          description: Reading category
        - name: provider
          in: query
          required: false
          schema:
            type: string
            enum: [claude, deepseek, gemini, openai]
          description: Preferred AI provider (optional, uses configured default if not specified)
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: User ID for subscription check
      responses:
        '200':
          description: Reading generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingResponse'
        '400':
          description: Invalid category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Premium subscription required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Chart not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/reading/{chartId}/chat:
    post:
      summary: Ask a follow-up question
      operationId: askQuestion
      tags:
        - Reading
      parameters:
        - name: chartId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Chart ID
        - name: provider
          in: query
          required: false
          schema:
            type: string
            enum: [claude, deepseek, gemini, openai]
          description: Preferred AI provider (optional, uses configured default if not specified)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Question answered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '404':
          description: Chart not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/reading/{chartId}/cached:
    get:
      summary: Get cached readings for a chart
      operationId: getCachedReadings
      tags:
        - Reading
      parameters:
        - name: chartId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Chart ID
      responses:
        '200':
          description: Cached readings list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CachedReadingsResponse'
  /api/reading/{chartId}/chat/history:
    get:
      summary: Get chat history for a chart
      operationId: getChatHistory
      tags:
        - Reading
      parameters:
        - name: chartId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Chart ID
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: User ID
      responses:
        '200':
          description: Chat history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatHistoryResponse'
        '404':
          description: Chart not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Clear chat history for a chart
      operationId: clearChatHistory
      tags:
        - Reading
      parameters:
        - name: chartId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Chart ID
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: User ID
      responses:
        '200':
          description: Chat history cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '404':
          description: Chart not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/subscription/status:
    get:
      summary: Get subscription status
      operationId: getSubscriptionStatus
      tags:
        - Subscription
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: User ID (optional for authenticated requests)
      responses:
        '200':
          description: Subscription status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionStatusResponse'
...
  /api/subscription/question-usage:
    get:
      summary: Get question usage for a user
      operationId: getQuestionUsage
      tags:
        - Subscription
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: User ID
      responses:
        '200':
          description: Question usage info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionUsageResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ChartResponse:
      type: object
      properties:
        success:
          type: boolean
        chart:
          $ref: '#/components/schemas/BirthChart'
    BirthChart:
      type: object
      required:
        - id
        - birthDate
        - birthTime
        - latitude
        - longitude
        - timezone
        - ascendant
        - planets
        - houses
        - dashas
        - ayanamsa
        - ayanamsaName
        - calculatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Name of the person
        birthDate:
          type: string
          format: date-time
        birthTime:
          type: string
        place:
          type: string
          description: Birth place name
        latitude:
          type: number
        longitude:
          type: number
        timezone:
          type: string
        ascendant:
          $ref: '#/components/schemas/Planet'
        planets:
          type: array
          items:
            $ref: '#/components/schemas/Planet'
        houses:
          type: array
          items:
            $ref: '#/components/schemas/House'
        dashas:
          type: array
          items:
            $ref: '#/components/schemas/DashaPeriod'
        ayanamsa:
          type: number
          description: Ayanamsa value in degrees
        ayanamsaName:
          type: string
          example: Lahiri
        calculatedAt:
          type: string
          format: date-time
    Planet:
      type: object
      required:
        - name
        - sign
        - signIndex
        - degree
        - longitude
        - nakshatra
        - nakshatraPada
        - house
        - isRetrograde
      properties:
        name:
          type: string
          enum: [Sun, Moon, Mars, Mercury, Jupiter, Venus, Saturn, Rahu, Ketu]
        sign:
          type: string
          enum: [Aries, Taurus, Gemini, Cancer, Leo, Virgo, Libra, Scorpio, Sagittarius, Capricorn, Aquarius, Pisces]
        signIndex:
          type: integer
          minimum: 0
          maximum: 11
        degree:
          type: number
          minimum: 0
          maximum: 30
        longitude:
          type: number
          minimum: 0
          maximum: 360
        nakshatra:
          type: string
        nakshatraPada:
          type: integer
          minimum: 1
          maximum: 4
        house:
          type: integer
          minimum: 1
          maximum: 12
        isRetrograde:
          type: boolean
    House:
      type: object
      required:
        - number
        - sign
        - signIndex
        - degree
        - longitude
      properties:
        number:
          type: integer
          minimum: 1
          maximum: 12
        sign:
          type: string
        signIndex:
          type: integer
        degree:
          type: number
        longitude:
          type: number
    DashaPeriod:
      type: object
      required:
        - planet
        - startDate
        - endDate
        - level
      properties:
        planet:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        level:
          type: string
          enum: [maha, antar, pratyantar]

    # Reading Schemas
    CategoriesResponse:
      type: object
      properties:
        success:
          type: boolean
        categories:
          type: array
          items:
            $ref: '#/components/schemas/ReadingCategory'
        userTier:
          type: string
          enum: [free, premium, pro]
        isPremium:
          type: boolean
    ReadingCategory:
      type: object
      properties:
        id:
          type: string
          enum: [summary, love, career, finances, health, timeline]
        name:
          type: string
        description:
          type: string
        accessLevel:
          type: string
          enum: [free, premium]
        isLocked:
          type: boolean
    ReadingResponse:
      type: object
      properties:
        success:
          type: boolean
        reading:
          $ref: '#/components/schemas/Reading'
    Reading:
      type: object
      required:
        - category
        - content
        - cached
        - chartId
        - generatedAt
      properties:
        category:
          type: string
        content:
          type: string
        cached:
          type: boolean
        chartId:
          type: string
          format: uuid
        generatedAt:
          type: string
          format: date-time
        provider:
          type: string
          description: AI provider used to generate the reading
          enum: [claude, deepseek, gemini, openai]
    ChatRequest:
      type: object
      required:
        - question
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 500
        history:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessageItem'
        userId:
          type: string
          description: User ID for tracking question quota
    ChatMessageItem:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        timestamp:
          type: string
          format: date-time
    ChatResponse:
      type: object
      properties:
        success:
          type: boolean
        response:
          type: string
        chartId:
          type: string
        timestamp:
          type: string
          format: date-time
        provider:
          type: string
          description: AI provider used to generate the response
          enum: [claude, deepseek, gemini, openai]
        questionsRemaining:
          type: integer
          description: Number of free questions remaining for the user
    QuestionQuotaResponse:
      type: object
      properties:
        success:
          type: boolean
        questionsUsed:
          type: integer
        questionsRemaining:
          type: integer
        limit:
          type: integer
        resetAt:
          type: string
          format: date-time
    ChatHistoryResponse:
      type: object
      properties:
        success:
          type: boolean
        chartId:
          type: string
        history:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessageItem'
        messageCount:
          type: integer
    CachedReadingsResponse:
      type: object
      properties:
        success:
          type: boolean
        chartId:
          type: string
        cachedReadings:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              createdAt:
                type: string
                format: date-time
              hasContent:
                type: boolean

    # AI Schemas
...
    SubscriptionStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        subscription:
          type: object
          properties:
            isPremium:
              type: boolean
            tier:
              type: string
              enum: [free, premium, pro]
            expiresAt:
              type: string
              format: date-time
              nullable: true
            provider:
              type: string
              nullable: true
              description: Subscription provider (revenuecat, stripe, manual) or null if free
            entitlements:
              type: array
              items:
                type: string
              description: List of entitlements (premium_access, unlimited_questions, all_readings)
...
    QuestionUsageResponse:
      type: object
      properties:
        success:
          type: boolean
        usage:
          type: object
          properties:
            count:
              type: integer
            remaining:
              type: integer
            limit:
              type: integer
            resetAt:
              type: string
              format: date-time
            isPremium:
              type: boolean

    # Error Schema
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
        path:
          type: string